+++ Tuki-Backend-Provisional/src/main/java/com/Tuki/Tuki_Backend_Provisional/Entidades/DTOs/ProductoDTOs/ProductoRespuestaDTO.java
@@ -6,7 +6,7 @@ public record ProductoRespuestaDTO (
         Long id,
         String nombre,
         double precio,
-        Long stock,
+        Integer stock,
         String urlImagen,
         String categoria
 ){}
