+++ Tuki-Backend-Provisional/src/main/java/com/Tuki/Tuki_Backend_Provisional/Entidades/DTOs/ProductoDTOs/ProductoUpdateDTO.java
@@ -7,7 +7,7 @@ public record ProductoUpdateDTO(
         String nombre,
         String descripcion,
         Double precio,
-        Long stock,
+        Integer stock,
         String urlImagen,
         Long categoriaId
 ){}
